refactor(AdminReservationApiDocs): 문서에 필요한 annotation interface로 분리
