+++ src/main/java/roomescape/controller/api/docs/AdminReservationApiDocs.java
@@ -1,18 +1,8 @@
-package roomescape.controller.api;
+package roomescape.controller.api.docs;
 
-import java.net.URI;
 import java.util.List;
 
-import jakarta.validation.Valid;
-
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -24,19 +14,9 @@ import roomescape.controller.dto.request.CreateReservationRequest;
 import roomescape.controller.dto.request.SearchReservationFilterRequest;
 import roomescape.controller.dto.response.ErrorMessageResponse;
 import roomescape.controller.dto.response.ReservationResponse;
-import roomescape.global.exception.RoomescapeException;
-import roomescape.service.AdminReservationService;
 
 @Tag(name = "AdminReservation", description = "관리자 예약 관련 API")
-@RestController
-@RequestMapping("/admin/reservations")
-public class AdminReservationController {
-    private final AdminReservationService adminReservationService;
-
-    public AdminReservationController(AdminReservationService adminReservationService) {
-        this.adminReservationService = adminReservationService;
-    }
-
+public interface AdminReservationApiDocs {
     @Operation(summary = "모든 예약 조회", description = "모든 사용자의 예약을 조회할 수 있다.")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "조회 성공",
@@ -46,11 +26,7 @@ public class AdminReservationController {
                     content = {@Content(
                             mediaType = "application/json", schema = @Schema(implementation = ErrorMessageResponse.class))}),
     })
-    @GetMapping
-    public ResponseEntity<List<ReservationResponse>> findAll() {
-        List<ReservationResponse> response = adminReservationService.findAllReserved();
-        return ResponseEntity.ok(response);
-    }
+    ResponseEntity<List<ReservationResponse>> findAll();
 
     @Operation(summary = "특정 예약 조회", description = "해당 기간에 해당 테마, 멤버인 예약들을 조회할 수 있다.")
     @ApiResponses(value = {
@@ -64,11 +40,7 @@ public class AdminReservationController {
                     content = {@Content(
                             mediaType = "application/json", schema = @Schema(implementation = ErrorMessageResponse.class))}),
     })
-    @GetMapping("/search")
-    public ResponseEntity<List<ReservationResponse>> find(SearchReservationFilterRequest request) {
-        List<ReservationResponse> response = adminReservationService.findAllByFilter(request);
-        return ResponseEntity.ok(response);
-    }
+    ResponseEntity<List<ReservationResponse>> find(SearchReservationFilterRequest request);
 
     @Operation(summary = "모든 예약 대기 조회", description = "모든 사용자의 예약 대기을 조회할 수 있다.")
     @ApiResponses(value = {
@@ -79,11 +51,7 @@ public class AdminReservationController {
                     content = {@Content(
                             mediaType = "application/json", schema = @Schema(implementation = ErrorMessageResponse.class))}),
     })
-    @GetMapping("/standby")
-    public ResponseEntity<List<ReservationResponse>> findAllStandby() {
-        List<ReservationResponse> response = adminReservationService.findAllStandby();
-        return ResponseEntity.ok(response);
-    }
+    ResponseEntity<List<ReservationResponse>> findAllStandby();
 
     @Operation(summary = "예약 생성", description = "예약을 생성할 수 있다.")
     @ApiResponses(value = {
@@ -97,13 +65,7 @@ public class AdminReservationController {
                     content = {@Content(
                             mediaType = "application/json", schema = @Schema(implementation = ErrorMessageResponse.class))}),
     })
-    @PostMapping
-    public ResponseEntity<ReservationResponse> save(@Valid @RequestBody CreateReservationRequest request) {
-        ReservationResponse response = adminReservationService.reserve(request);
-
-        return ResponseEntity.created(URI.create("/reservations/" + response.id()))
-                .body(response);
-    }
+    ResponseEntity<ReservationResponse> save(CreateReservationRequest request);
 
     @Operation(summary = "예약 삭제", description = "예약을 삭제할 수 있다.")
     @ApiResponses(value = {
@@ -116,11 +78,7 @@ public class AdminReservationController {
                     content = {@Content(
                             mediaType = "application/json", schema = @Schema(implementation = ErrorMessageResponse.class))}),
     })
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
-        adminReservationService.deleteById(id);
-        return ResponseEntity.noContent().build();
-    }
+    ResponseEntity<Void> delete(Long id);
 
     @Operation(summary = "예약 대기 삭제", description = "예약 대기를 삭제할 수 있다.")
     @ApiResponses(value = {
@@ -133,9 +91,5 @@ public class AdminReservationController {
                     content = {@Content(
                             mediaType = "application/json", schema = @Schema(implementation = ErrorMessageResponse.class))}),
     })
-    @DeleteMapping("/standby/{id}")
-    public ResponseEntity<Void> deleteStandby(@PathVariable Long id) {
-        adminReservationService.deleteStandby(id);
-        return ResponseEntity.noContent().build();
-    }
+    ResponseEntity<Void> deleteStandby(Long id);
 }
